+++ qcengine/programs/molpro.py
@@ -45,11 +45,21 @@ class MolproHarness(ProgramHarness):
         "M12HFC", "HJSWPBE", "HJSWPBEH", "TCSWPBE", "PBESOL"
     }
 
-    # Currently supported methods in QCEngine for Molpro
-    _scf_methods: Set[str] = {"HF", "RHF", "KS", "RKS"}
-    _post_hf_methods: Set[str] = {'MP2', 'CCSD', 'CCSD(T)'}
-    _supported_methods: Set[str] = {*_scf_methods, *_post_hf_methods}
-
+    # Different open-shell scenarios:
+    # - Restricted references
+    #       RHF-RMP2 (like RHF-UMP2 using Molpro's naming convention for CC methods)
+    #       RHF-UCCSD
+    #       RHF-UCCSD(T)
+    #       RHF-RCCSD
+    #       RHF-RCCSD(T)
+    # - Unrestricted references (Only supported up to UMP2, no CC support)
+    #       UHF-UMP2
+    # NOTE: Unrestricted SCF methods must be specified by using keyword reference
+    _hf_methods: Set[str] = {"HF", "RHF"}
+    _restricted_post_hf_methods: Set[str] = {"MP2", "CCSD", "CCSD(T)"}  # RMP2, RCCSD, RCCSD(T)}
+    # TODO Add keyword to specify unrestricted for WF method
+    # _unrestricted_post_hf_methods: Set[str] = {"UMP2", "UCCSD", "UCCSD(T)"}
+    _post_hf_methods: Set[str] = {*_restricted_post_hf_methods}
 
     class Config(ProgramHarness.Config):
         pass
@@ -151,12 +161,16 @@ class MolproHarness(ProgramHarness):
         if template is None:
             input_file = []
 
+            # Resolving keywords
+            caseless_keywords = {k.lower(): v for k, v in input_model.keywords.items()}
+            unrestricted = False
+            # Following Molpro default that ROHF is done for open-shell calculations unless unrestricted is specified
+            if "reference" in caseless_keywords and caseless_keywords["reference"] == "unrestricted":
+                unrestricted = True
+
             # Memory is in megawords per core for Molpro
             memory_mw_core = int(config.memory * (1024**3) / 8e6 / config.ncores)
             input_file.append("memory,{},M".format(memory_mw_core))
-            # TODO Decide how I want this keyword to look/work
-            # if input_model.keywords.get('wfu') is not None:
-            #     input_file.append('file,2,{}'.format(input_model.keywords.get('wfu')))
             input_file.append('')
 
             # Write the geom
@@ -169,16 +183,27 @@ class MolproHarness(ProgramHarness):
             input_file.append('}')
             input_file.append('')
 
+            # Determine what SCF type (restricted vs. unrestricted)
+            hf_type = 'RHF'
+            dft_type = 'RKS'
+            if unrestricted:
+                hf_type = 'UHF'
+                dft_type = 'UKS'
+
             # Write energy call
             energy_call = []
             # If post-hf method is called then make sure to write a HF call first
-            if input_model.model.method.upper() in self._post_hf_methods:
-                energy_call.append('{HF}')
-            # If DFT call make sure to write {rks,method}
-            if input_model.model.method.upper() in self._dft_functionals:
-                energy_call.append(f'{{rks,{input_model.model.method}}}')
-            else:
+            if input_model.model.method.upper() in self._post_hf_methods:  # post SCF case
+                energy_call.append(f'{{{hf_type}}}')
+                energy_call.append('')
                 energy_call.append(f'{{{input_model.model.method}}}')
+            # If DFT call make sure to write {rks,method}
+            elif input_model.model.method.upper() in self._dft_functionals:  # DFT case
+                energy_call.append(f'{{{dft_type},{input_model.model.method}}}')
+            elif input_model.model.method.upper() in self._hf_methods:  # HF case
+                energy_call.append(f'{{{hf_type}}}')
+            else:
+                raise InputError(f'Method {input_model.model.method} not implemented for Molpro.')
 
             # Write appropriate driver call
             if input_model.driver == 'energy':
@@ -229,6 +254,7 @@ class MolproHarness(ProgramHarness):
         #      - orbitals
         output_data = {}
         properties = {}
+        extras = {}
         name_space = {'molpro_uri': 'http://www.molpro.net/schema/molpro-output'}
 
         # Molpro commands map
@@ -236,8 +262,10 @@ class MolproHarness(ProgramHarness):
             "Energy": {
                 "HF": "scf_total_energy",
                 "RHF": "scf_total_energy",
+                "UHF": "scf_total_energy",
                 "KS": "scf_total_energy",
-                "RKS": "scf_total_energy"
+                "RKS": "scf_total_energy",
+                "UKS": "scf_total_energy"
             },
             "total energy": {
                 "MP2": "mp2_total_energy",
@@ -261,56 +289,76 @@ class MolproHarness(ProgramHarness):
             "Dipole moment": {
                 "HF": "scf_dipole_moment",
                 "RHF": "scf_dipole_moment",
+                "UHF": "scf_dipole_moment",
                 "KS": "scf_dipole_moment",
                 "RKS": "scf_dipole_moment",
+                "UKS": "scf_dipole_moment",
                 "MP2": "mp2_dipole_moment",
                 "CCSD": "ccsd_dipole_moment",
                 "CCSD(T)": "ccsd_prt_pr_dipole_moment"
             }
         }
 
+        # Started adding basic support for local correlation methods in Molpro
+        molpro_extras_map = {
+            "total energy": {
+                "LMP2": "local_mp2_total_energy",
+                "LCCSD": "local_ccsd_total_energy",
+                "LCCSD(T0)": "local_ccsd_prt0_pr_total_energy",
+                "LCCSD(T)": "local_ccsd_prt_pr_total_energy"
+            },
+            "correlation energy": {
+                "LMP2": "local_mp2_correlation_energy",
+                "LCCSD": "local_ccsd_correlation_energy"
+            },
+            "singlet pair energy": {
+                "LMP2": "local_mp2_singlet_pair_energy",
+                "LCCSD": "local_ccsd_singlet_pair_energy"
+            },
+            "triplet pair energy": {
+                "LMP2": "local_mp2_triplet_pair_energy",
+                "LCCSD": "local_ccsd_triplet_pair_energy"
+            },
+            "singles energy": {
+                "LCCSD": "local_ccsd_singles_energy"
+            },
+            # "strong pair energy": {
+            #     "LCCSD": "local_ccsd_strong_pair_energy"
+            # },
+            # "weak pair energy": {
+            #     "LMP2": "local_mp2_weak_pair_energy"
+            # }
+        }
+
         # Molpro variables map used for quantities not found in the command map
-        molpro_var_map = {
+        molpro_variable_map = {
             "_ENUC": "nuclear_repulsion_energy",
-            "_DFTFUN": "scf_xc_energy"
+            "_DFTFUN": "scf_xc_energy",
+            "_NELEC": ["calcinfo_nalpha", "calcinfo_nbeta"]
             # "_EMP2_SCS": "scs_mp2_total_energy"
         }
 
+        # Process data in molpro_map
         # Loop through each jobstep
-        # The jobstep tag in Molpro contains output from commands (e.g. {hf}, {force})
+        # The jobstep tag in Molpro contains output from commands (e.g. {HF}, {force})
         for jobstep in root.findall('molpro_uri:job/molpro_uri:jobstep', name_space):
-            # Remove the -SCF part of the command string when Molpro calls HF or KS
             command = jobstep.attrib['command']
-            if '-SCF' in command:
-                command = command[:-4]
-            # Grab energies and dipole moment
-            if command in self._supported_methods:
-                for child in jobstep.findall('molpro_uri:property', name_space):
-                    prop_name = child.attrib['name']
-                    prop_method = child.attrib['method']
-                    value = child.attrib['value']
-                    if prop_name in molpro_map:
-                        if prop_method in molpro_map[prop_name]:
-                            if prop_name == "Dipole moment":
-                                properties[molpro_map[prop_name][prop_method]] = [float(x) for x in value.split()]
-                            else:
-                                properties[molpro_map[prop_name][prop_method]] = float(value)
-            # Grab gradient
-            elif 'FORCE' in command:
+            if 'FORCE' in command:  # Grab gradient
                 for child in jobstep.findall('molpro_uri:gradient', name_space):
                     # Stores gradient as a single list where the ordering is [1x, 1y, 1z, 2x, 2y, 2z, ...]
                     output_data['return_result'] = [float(x) for x in child.text.split()]
-
-        # Look for final energy in the molecule tag in case it's needed
-        # Note: For the DFT case mol_method is the name of the functional plus R or U in front
-        molecule = root.find('molpro_uri:job/molpro_uri:molecule', name_space)
-        mol_method = molecule.attrib['method']
-        mol_final_energy = float(molecule.attrib['energy'])
-        # Loop over each variable under the variables tag to grab additional info from molpro_var_map
-        for variable in molecule.findall('molpro_uri:variables/molpro_uri:variable', name_space):
-            var_name = variable.attrib['name']
-            if var_name in molpro_var_map:
-                properties[molpro_var_map[var_name]] = float(variable[0].text)
+            else:  # Grab energies and dipole moment
+                for child in jobstep.findall('molpro_uri:property', name_space):
+                    property_name = child.attrib['name']
+                    property_method = child.attrib['method']
+                    value = child.attrib['value']
+                    if property_name in molpro_map and property_method in molpro_map[property_name]:
+                        if property_name == "Dipole moment":
+                            properties[molpro_map[property_name][property_method]] = [float(x) for x in value.split()]
+                        else:
+                            properties[molpro_map[property_name][property_method]] = float(value)
+                    elif property_name in molpro_extras_map and property_method in molpro_extras_map[property_name]:
+                        extras[molpro_extras_map[property_name][property_method]] = float(value)
 
         # Convert triplet and singlet pair correlation energies to opposite-spin and same-spin correlation energies
         if 'mp2_singlet_pair_energy' in properties and 'mp2_triplet_pair_energy' in properties:
@@ -329,37 +377,64 @@ class MolproHarness(ProgramHarness):
             del properties['ccsd_singlet_pair_energy']
             del properties['ccsd_triplet_pair_energy']
 
-        # Grab the method from input
-        method = input_model.model.method
-        if method.upper() in self._dft_functionals:  # Determine if method is a DFT functional
-            method = "RKS"
+        # Process data in molpro_variable_map
+        # Note: For the DFT case molecule_method is the name of the functional plus R or U in front
+        molecule = root.find('molpro_uri:job/molpro_uri:molecule', name_space)
+        molecule_method = molecule.attrib['method']
+        molecule_final_energy = float(molecule.attrib['energy'])  # Energy from the molecule tag in case its needed
+        # Loop over each variable under the variables tag to grab additional info from molpro_variable_map
+        for variable in molecule.findall('molpro_uri:variables/molpro_uri:variable', name_space):
+            variable_name = variable.attrib['name']
+            if variable_name in molpro_variable_map:
+                if variable_name == "_NELEC":
+                    nelec = int(float(variable[0].text))
+                    nunpaired = (input_model.molecule.molecular_multiplicity - 1)
+                    nbeta = (nelec - nunpaired) // 2
+                    nalpha = nelec - nbeta
+                    properties[molpro_variable_map[variable_name][0]] = nalpha
+                    properties[molpro_variable_map[variable_name][1]] = nbeta
+                else:
+                    properties[molpro_variable_map[variable_name]] = float(variable[0].text)
 
-        # A slightly more robust way of determining the final energy.
+        # Process basis set data
+        basis_set = root.find('molpro_uri:job/molpro_uri:molecule/molpro_uri:basisSet', name_space)
+        nbasis = int(basis_set.attrib['length'])
+        # angular_type = basis_set.attrib['angular']  # cartesian vs spherical
+        properties["calcinfo_nbasis"] = nbasis
+
+        # Grab the method from input
+        method = input_model.model.method.upper()
+
+        # Determining the final energy
         # Throws an error if the energy isn't found for the method specified from the input_model.
-        if method in self._post_hf_methods and molpro_map['total energy'][method] in properties:
+        if method in molpro_map['total energy'].keys() and molpro_map['total energy'][method] in properties:
             final_energy = properties[molpro_map['total energy'][method]]
-        elif method in self._scf_methods and molpro_map['Energy'][method] in properties:
+        elif method in molpro_map['Energy'].keys() and molpro_map['Energy'][method] in properties:
             final_energy = properties[molpro_map['Energy'][method]]
         else:
             # Back up method for determining final energy if not already present in properties
             # Use the total energy from the molecule tag if it matches the input method
-            if input_model.model.method in mol_method:
-                final_energy = mol_final_energy
+            # if input_model.model.method.upper() in molecule_method:
+            if method in molecule_method:
+                final_energy = molecule_final_energy
                 if method in self._post_hf_methods:
-                    properties[molpro_map['total energy'][method]] = mol_final_energy
+                    properties[molpro_map['total energy'][method]] = molecule_final_energy
                     properties[molpro_map['correlation energy']
-                               [method]] = mol_final_energy - properties['scf_total_energy']
-                elif method in self._scf_methods:
-                    properties[molpro_map['Energy'][method]] = mol_final_energy
+                               [method]] = molecule_final_energy - properties['scf_total_energy']
+                elif method in self._dft_functionals:
+                    properties[molpro_map['Energy']["KS"]] = molecule_final_energy
+                elif method in self._hf_methods:
+                    properties[molpro_map['Energy'][method]] = molecule_final_energy
             else:
                 raise KeyError(f"Could not find {method} total energy")
 
-        # Replace return_result with final_energy if gradient wasn't called
+        # Determining return_result
         if "return_result" not in output_data:
             output_data["return_result"] = final_energy
 
         # Final output_data assignments needed for the Result object
         output_data["properties"] = properties
+        output_data["extras"] = extras
         output_data['schema_name'] = 'qcschema_output'
         output_data['stdout'] = outfiles["dispatch.out"]
         output_data['success'] = True